backend/requirements.txt
Flask
python-dotenv
pymongo
sentence_transformers
flask-cors
pydantic==2.7.4
openai==1.35.3
chromadb
